refactored extension methods around the elementmodel to now be in the "appropriate" classes
