+++ src/Hl7.Fhir.R4/CompatibilitySuppressions.xml
@@ -3,9 +3,23 @@
 <Suppressions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
   <Suppression>
     <DiagnosticId>CP0001</DiagnosticId>
+    <Target>T:Hl7.Fhir.ElementModel.PocoBuilderExtensions</Target>
+    <Left>lib/net8.0/Hl7.Fhir.R4.dll</Left>
+    <Right>lib/net8.0/Hl7.Fhir.R4.dll</Right>
+    <IsBaselineSuppression>true</IsBaselineSuppression>
+  </Suppression>
+  <Suppression>
+    <DiagnosticId>CP0001</DiagnosticId>
     <Target>T:Hl7.Fhir.ElementModel.TypedElementExtensions</Target>
     <Left>lib/net8.0/Hl7.Fhir.R4.dll</Left>
     <Right>lib/net8.0/Hl7.Fhir.R4.dll</Right>
+    <IsBaselineSuppression>true</IsBaselineSuppression>
+  </Suppression>
+  <Suppression>
+    <DiagnosticId>CP0001</DiagnosticId>
+    <Target>T:Hl7.Fhir.ElementModel.PocoBuilderExtensions</Target>
+    <Left>lib/netstandard2.1/Hl7.Fhir.R4.dll</Left>
+    <Right>lib/netstandard2.1/Hl7.Fhir.R4.dll</Right>
     <IsBaselineSuppression>true</IsBaselineSuppression>
   </Suppression>
   <Suppression>
