+++ src/Hl7.Fhir.Shims.STU3AndUp/Hl7.Fhir.Shims.STU3AndUp.projitems
@@ -9,8 +9,7 @@
     <Import_RootNamespace>Hl7.Fhir.Shims.STU3AndUp</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)ElementModel\PocoBuilderExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ElementModel\TypedElementExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ElementModel\VersionedConversionExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FhirPath\FhirPathExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\HumanName.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Model\Initializers.cs" />
